Add transparent UTXOs to `z_listunspent` outputs.
This merges the functionality of the zcashd `listunspent` RPC method
with the listing of shielded UTXOs already provided by `z_listunspent`.
It updates the RPC API of `z_listunspent` to include all of the
arguments supported by this API in `zcashd`, although at present the
`include_watch_only`, `maxconf`, and `addresses` argument values are
ignored.