rpc (z_sendmany): Build transaction proposal synchronously.
The primary purpose of the asynchronous operation of `z_sendmany` is to
avoid long proving times in a synchronous call. We don't have any reason
to believe that proposal construction should be a long-running process,
and it's better to report errors to the user synchronously rather than
async, so this change moves all of proposal construction into the
synchronous phase of the RPC call.