z_viewtransaction: Temporarily omit block data if we haven't scanned it
We cannot blindly query the chain for a height from the wallet, as a
reorg might have occurred in the meantime. Instead we need to wait until
the wallet knows the block hash corresponding to its view of the
transaction's mined height, and then fetch that. But if the wallet
starts returning the mined height as fetched from the UTXO set from
`WalletRead::get_tx_height` (instead of it currently querying the blocks
table) then we won't have the block data. But we *do* want it to return
this value so that `status` and `confirmations` are set consistent with
the wallet's view of the chain.

This will all be resolved by a migration to Zaino's `ChainIndex` API so
for now we just omit the details we can't obtain.